Add better search summary
