+++ cosmetics-web/app/helpers/application_helper.rb
@@ -35,16 +35,4 @@ module ApplicationHelper
 
     govukErrorSummary(titleText: "There is a problem", errorList: error_list)
   end
-
-  def display_keywords(keywords)
-    return if keywords.blank?
-
-    keywords = keywords.split(" ")
-    keywords = keywords.join(", ")
-    keywords = keywords.gsub(",,", ",")
-
-    "<span class=\"govuk-!-font-weight-bold\">#{keywords},</span>".html_safe
-  rescue StandardError
-    keywords
-  end
 end
