+++ cosmetics-web/app/views/poison_centres/notifications/_date_filter_input.html.erb
@@ -1,5 +1,5 @@
-<% if @form.errors[attribute].present? %>
-  <% @form.errors[attribute].each do |error| %>
+<% if @search_form.errors[attribute].present? %>
+  <% @search_form.errors[attribute].each do |error| %>
     <span class="govuk-error-message">
       <span class="govuk-visually-hidden">Error:</span> <%= error %>
     </span>
@@ -12,20 +12,20 @@
       <label class="govuk-label govuk-date-input__label" for="notification_search_form_<%= attribute %>_day">
         Day
       </label>
-      <input class="govuk-input govuk-date-input__input govuk-input--width-2 <%= @form.errors[attribute].present? ? 'govuk-input--error' : '' %>" id="notification_search_form_<%= attribute %>_day" value="<%= @form.public_send(attribute)&.day %>" name="notification_search_form[<%= attribute %>][day]" type="text" pattern="[0-9]*" inputmode="numeric"></div>
+      <input class="govuk-input govuk-date-input__input govuk-input--width-2 <%= @search_form.errors[attribute].present? ? 'govuk-input--error' : '' %>" id="notification_search_form_<%= attribute %>_day" value="<%= @search_form.public_send(attribute)&.day %>" name="notification_search_form[<%= attribute %>][day]" type="text" pattern="[0-9]*" inputmode="numeric"></div>
   </div>
   <div class="govuk-date-input__item">
     <div class="govuk-form-group">
       <label class="govuk-label govuk-date-input__label" for="notification_search_form_<%= attribute %>_month">
         Month
       </label>
-      <input class="govuk-input govuk-date-input__input govuk-input--width-2 <%= @form.errors[attribute].present? ? 'govuk-input--error' : '' %>" id="notification_search_form_<%= attribute %>_month" value="<%= @form.public_send(attribute)&.month %>" name="notification_search_form[<%= attribute %>][month]" type="text" pattern="[0-9]*" inputmode="numeric"></div>
+      <input class="govuk-input govuk-date-input__input govuk-input--width-2 <%= @search_form.errors[attribute].present? ? 'govuk-input--error' : '' %>" id="notification_search_form_<%= attribute %>_month" value="<%= @search_form.public_send(attribute)&.month %>" name="notification_search_form[<%= attribute %>][month]" type="text" pattern="[0-9]*" inputmode="numeric"></div>
   </div>
   <div class="govuk-date-input__item">
     <div class="govuk-form-group">
       <label class="govuk-label govuk-date-input__label" for="notification_search_form_<%= attribute %>_year">
         Year
       </label>
-      <input class="govuk-input govuk-date-input__input govuk-input--width-4 <%= @form.errors[attribute].present? ? 'govuk-input--error' : '' %>" id="notification_search_form_<%= attribute %>_year" value="<%= @form.public_send(attribute)&.year %>" name="notification_search_form[<%= attribute %>][year]" type="text" pattern="[0-9]*" inputmode="numeric"></div>
+      <input class="govuk-input govuk-date-input__input govuk-input--width-4 <%= @search_form.errors[attribute].present? ? 'govuk-input--error' : '' %>" id="notification_search_form_<%= attribute %>_year" value="<%= @search_form.public_send(attribute)&.year %>" name="notification_search_form[<%= attribute %>][year]" type="text" pattern="[0-9]*" inputmode="numeric"></div>
   </div>
 </div>
