+++ cosmetics-web/app/views/poison_centres/notifications/_hidden_date_fields.html.erb
@@ -1,3 +1,3 @@
-<input id="notification_search_form_<%= attribute %>_day_hidden" value="<%= @form.public_send(attribute)&.day %>" name="notification_search_form[<%= attribute %>][day]" type="hidden">
-<input id="notification_search_form_<%= attribute %>_month_hidden" value="<%= @form.public_send(attribute)&.month %>" name="notification_search_form[<%= attribute %>][month]" type="hidden">
-<input id="notification_search_form_<%= attribute %>_year_hidden" value="<%= @form.public_send(attribute)&.year %>" name="notification_search_form[<%= attribute %>][year]" type="hidden">
+<input id="notification_search_form_<%= attribute %>_day_hidden" value="<%= @search_form.public_send(attribute)&.day %>" name="notification_search_form[<%= attribute %>][day]" type="hidden">
+<input id="notification_search_form_<%= attribute %>_month_hidden" value="<%= @search_form.public_send(attribute)&.month %>" name="notification_search_form[<%= attribute %>][month]" type="hidden">
+<input id="notification_search_form_<%= attribute %>_year_hidden" value="<%= @search_form.public_send(attribute)&.year %>" name="notification_search_form[<%= attribute %>][year]" type="hidden">
