+++ cosmetics-web/app/views/poison_centres/notifications/index.html.erb
@@ -2,9 +2,9 @@
 
 <div class="govuk-grid-row">
   <section class="govuk-grid-column-full">
-    <%= error_summary(@form.errors) %>
+    <%= error_summary(@search_form.errors) %>
 
-    <%= form_for(@form, url: poison_centre_notifications_path, method: :get, html: { 'novalidate' => true }) do |form| %>
+    <%= form_for(@search_form, url: poison_centre_notifications_path, method: :get, html: { 'novalidate' => true }) do |form| %>
       <div class="govuk-form-group">
         <h1 class="govuk-label-wrapper">
           <label class="govuk-label govuk-label--l" for="notification_search_form_q">
